Database server changed - home
ApplicationDbContext hotfix.

+++ Api/Api.DAL/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -28,8 +28,8 @@ namespace Api.DAL.Migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("Price")
-                        .HasColumnType("int");
+                    b.Property<long>("Price")
+                        .HasColumnType("bigint");
 
                     b.HasKey("Name");
